requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
dnspython==2.3.0
eventlet==0.33.3
Flask==2.3.2
Flask-SocketIO==5.3.4
frozenlist==1.4.0
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.23.6
python-engineio==4.5.1
python-socketio==5.8.0
requests==2.31.0
six==1.16.0
tqdm==4.65.0
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
tika==2.6.0
flask-cors==4.0.0
python-dotenv==1.0.1